# specs/swagger.yml
openapi: 3.0.0
info:
  description: |
    Auth Microservice

    [link to repo](https://<NASCO_BITBUCKET_URL>/nodejs-template)
  version: '1.0'
  title: Auth Microservice
servers:
  - url: http://localhost:{port}/
    variables:
      port:
        default: "5050"
    description: local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
paths:
  /:
    get:
      tags:
        - default
      summary: Get the default route
      description: returns index
      operationId: default
      responses:
        200:
          description: successful api call, returns index
          content:
            application/vnd.nasco.auth: { }
        400:
          description: validation error
        500:
          description: an error occurred
...
  /account/verification/email:
    post:
      tags:
        - account
      summary: verify email
      description: verify email by validating code
      operationId: verifyEmail
      parameters:
        - in: header
          name: x-tenant-platform
          description: Set the tenant platform
          required: true
          schema:
            type: string
            enum:
              - NASCOEMIRATES
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  format: uuid
                  example: 6b96d1d4-3092-41a8-8e31-8bf0a1ba30a8
              required:
                - code
      responses:
        200:
          description: success
          content:
            application/vnd.nasco.auth: {}
        401:
          description: Unauthorized - Expired or invalid authorization token
          content:
            application/vnd.nasco.auth:
              schema:
                type: object
                properties:
                  status:
                    type: number
                  message:
                    type: string
              example:
                status: 401
                message: Unauthorized
        400:
          description: validation error / Invalid header
          content:
            application/vnd.nasco.auth:
              example:
                status: 400
                message: validation error
        500:
          description: an error occurred
          content:
            application/vnd.nasco.auth:
              example:
                status: 500
                message: Internal server error
  /account/recovery/password:
    post:
      tags:
        - account
      summary: used when a user has forgot his password and can't sign in
      description: Sends OTP by EMAIL or SMS according to the user"s choice. If user selects the SMS channel and has not set their mobile number yet, the API will throw a `406 - Not Acceptable` error
      operationId: requestPasswordRecovery
      parameters:
        - in: header
          name: x-tenant-platform
          description: Set the tenant platform
          required: true
          schema:
            type: string
            enum:
              - NASCOEMIRATES
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: john.doe
                isEmailChannel:
                  type: boolean
                  description: set this property to true if the user has a mobile number stored and yet prefers to receive the OTP by email
              required:
                - username
      responses:
        200:
          description: success
          content:
            application/vnd.nasco.auth: {}
        401:
          description: Unauthorized - Expired or invalid authorization token
          content:
            application/vnd.nasco.auth:
              schema:
                type: object
                properties:
                  status:
                    type: number
                  message:
                    type: string
              example:
                status: 401
                message: Unauthorized
        400:
          description: validation error / Invalid header
          content:
            application/vnd.nasco.auth:
              example:
                status: 400
                message: validation error
        500:
          description: an error occurred
          content:
            application/vnd.nasco.auth:
              example:
                status: 500
                message: Internal server error
...
  /account/recovery/password/reset:
    post:
      tags:
        - account
      summary: reset user password
      description: resets user password
      operationId: resetPassword
      parameters:
        - in: header
          name: x-tenant-platform
          description: Set the tenant platform
          required: true
          schema:
            type: string
            enum:
              - NASCOEMIRATES
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: john.doe
                newPassword:
                  type: string
                  format: password
                  example: myPassword
              required:
                - username
                - newPassword
      responses:
        200:
          description: success
          content:
            application/vnd.nasco.auth: {}
        401:
          description: Unauthorized - Expired or invalid authorization token
          content:
            application/vnd.nasco.auth:
              schema:
                type: object
                properties:
                  status:
                    type: number
                  message:
                    type: string
              example:
                status: 401
                message: Unauthorized
        400:
          description: validation error / Invalid header
          content:
            application/vnd.nasco.auth:
              example:
                status: 400
                message: validation error
        500:
          description: an error occurred
          content:
            application/vnd.nasco.auth:
              example:
                status: 500
                message: Internal server error
  /account/recovery/username:
    post:
      tags:
        - account
      summary: recover username by email
      description: Checks if email exists in the system, then sends an email including user"s username. If more than 1 account are linked to the same email, the username cannot be recovered (it will need additional info such as firstname, lastname, etc...)
      operationId: recoverUsername
      parameters:
        - in: header
          name: x-tenant-platform
          description: Set the tenant platform
          required: true
          schema:
            type: string
            enum:
              - NASCOEMIRATES
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: john.doe@example.com
              required:
                - email
      responses:
        200:
          description: success
          content:
            application/vnd.nasco.auth: {}
        401:
          description: Unauthorized - Expired or invalid authorization token
          content:
            application/vnd.nasco.auth:
              schema:
                type: object
                properties:
                  status:
                    type: number
                  message:
                    type: string
              example:
                status: 401
                message: Unauthorized
        400:
          description: validation error / Invalid header
          content:
            application/vnd.nasco.auth:
              example:
                status: 400
                message: validation error
        500:
          description: an error occurred
          content:
            application/vnd.nasco.auth:
              example:
                status: 500
                message: Internal server error